Added Game layout
Here is the layout for the game, more or less. It's not finalized, but it could probably be used to try out any code implementations and whatnot. Ignore the colors, I used them to hep with layout visualization. The important stuff is the active and passive text areas, the tabs, and the tab-content blocks