Finished the friend search algorithm
Here is the friend search algorithm. It should compile correctly, but I couldn't test it beforehand. There are a couple changes that need to be made before it will fit the code we have already written. For example, there is a threshold variable within the findFriendList function which controls the allowed percent similarity of the matched users. The query within the function must also be changed to query the correct table